pyproject.toml
[tool.poetry]
name = "aocref"
version = "2.0.1"
description = "Age of Empires reference data"
authors = ["SiegeEngineers Team <https://discord.com/invite/njAsNuD@discord.com>"]
repository = "https://github.com/SiegeEngineers/aoc-reference-data"
readme = "README.md"
packages = [
    { include = "ard_utility" },
    { include = "aocref" },
]
[tool.poetry.scripts]
ard_utility = "ard_utility.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.22.0"
"ruamel.yaml" = "^0.16.12"
pycountry = "^20.7.3"
docopt = "^0.6.2"

[tool.poetry.dev-dependencies]
poetry = "^1.1.12"
flake8 = "^4.0.1"
bandit = "^1.7.2"
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
